database/api/requirements.txt
fastapi>=0.115
uvicorn>=0.30
python-dotenv>=1.0
snowflake-connector-python>=3.10
pydantic>=2.8